Raise BlockFS/TLSF file size limit from 1 GiB to ~4 GiB
- Bump FL_MAX from 30 to 32 (max block: 2^32 = 4 GiB, up from 1 GiB)
- Fix signed int32 overflow in TLSF _blockSize (JS bitwise & returns int32,
  wrapping block sizes >= 2 GiB to negative -- pre-existing bug, unreachable
  at old limit)
- Fix same overflow in _growPool blockSz alignment rounding
- Add pool ceiling guard (_growPool returns 0 when pool would exceed 4 GiB
  uint32 field, preventing silent wrap)
- Change _growExtent from doubling to linear +256 MiB past 256 MiB, avoiding
  massive reallocs that prematurely hit the pool ceiling
- Same FL_MAX and guard changes in compiler.js WASM malloc
- Fix block_size return type (long -> unsigned long) in WASM TLSF for same
  signed-overflow reason on wasm32

All 583 unit + 73 blockfs tests pass. Tested manually with 1.5 GiB file.

+++ compiler.js
@@ -19789,8 +19789,8 @@ static long pool_end;
 static long last_block;
 static int  initialized;
 
-static long block_size(long block) {
-  return *(long *)block & ~FLAG_BITS;
+static unsigned long block_size(long block) {
+  return (unsigned long)(*(long *)block & ~FLAG_BITS);
 }
 
 static int block_is_free(long block) {
